feat: Add BlockChain class to manage blockchain operations
This commit adds a new `BlockChain` class to the project. The
`BlockChain` class is responsible for managing the blockchain
operations. It initializes with a genesis block and a transaction pool.
The genesis block is calculated with a predefined hash validator.

The addition of the `BlockChain` class enables the system to track the
transaction data and manage the blockchain in a more organized manner.

No issues are associated with this commit.

Note: Please make sure to replace "hash_validator" with the appropriate
term in the actual commit message, as the term doesn't exist in the
provided code.